requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
configparser==5.0.0
discord.py==1.3.4
idna==2.10
multidict==4.7.6
requests==2.24.0
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
